requirements.txt
Flask==3.0.3
Werkzeug==3.1.4
Jinja2==3.1.6
itsdangerous==2.2.0
click==8.3.1
blinker==1.9.0

python-dotenv==1.0.1
requests==2.32.5
urllib3==2.6.2
charset-normalizer==3.4.4
idna==3.11
certifi==2025.11.12

numpy==1.26.4
scipy==1.15.3
scikit-learn==1.7.2
joblib==1.5.3
threadpoolctl==3.6.0

faiss-cpu==1.8.0.post1
sentence-transformers==5.2.0
transformers==4.57.3
tokenizers==0.22.1
huggingface-hub==0.36.0
safetensors==0.7.0
regex==2025.11.3
tqdm==4.67.1

llama_cpp_python==0.2.90
diskcache==5.6.3
psutil==7.1.3
filelock==3.20.0
PyYAML==6.0.3
typing_extensions==4.15.0
